backend/pyproject.toml
[project]
name = "fraud-detection-backend"
version = "0.1.0"
description = "Multi-agent fraud detection system backend"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.2",
    "uvicorn[standard]>=0.27.1",
    "pydantic>=2.6.1",
    "pydantic-settings>=2.1.0",
    "python-dotenv>=1.0.1",
    "langgraph>=0.2.0",
    "langchain-core>=0.2.0",
    "langchain-openai>=0.1.1",
    "openai>=1.12.0",
    "chromadb>=0.4.22",
    "boto3>=1.34.34",
    "httpx>=0.26.0",
    "python-multipart>=0.0.9",
    "filelock>=3.13.1",
    "tiktoken>=0.6.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.uv]
dev-dependencies = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.2.0",
    "mypy>=1.8.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = false
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
